requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.7.4.post0=py36h7f8727e_2
altair=4.1.0=pypi_0
argon2-cffi=21.1.0=pypi_0
astor=0.8.1=pypi_0
async-generator=1.10=pypi_0
async-timeout=3.0.1=py36h06a4308_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
backports-zoneinfo=0.2.1=pypi_0
base58=2.1.1=pypi_0
beautifulsoup4=4.10.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
blinker=1.4=py36h06a4308_0
brotlipy=0.7.0=py36h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.3.29=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.5.30=py36h06a4308_0
cffi=1.15.0=pypi_0
chardet=4.0.0=py36h06a4308_1003
charset-normalizer=2.0.8=pypi_0
click=7.1.2=pypi_0
clip=1.0=pypi_0
coverage=5.5=py36h27cfd23_2
cryptography=3.4.7=py36hd23ed53_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.24=py36h295c915_0
dataclasses=0.8=pyh4f3eec9_6
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.3=py36_0
expat=2.4.1=h2531618_2
ffmpeg=4.2.2=h20bf706_0
filelock=3.4.0=pypi_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
ftfy=6.0.3=pypi_0
gdown=4.2.0=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.18=pypi_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.36.1=py36h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
huggingface-hub=0.4.0=pypi_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
idna_ssl=1.1.0=py36h06a4308_0
imageio=2.15.0=pypi_0
importlib-metadata=4.8.1=py36h06a4308_0
importlib-resources=5.4.0=pypi_0
importlib_metadata=1.5.0=py36_0
intel-openmp=2022.0.1=h06a4308_3633
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pypi_0
jedi=0.17.0=py36_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=pypi_0
jupyter_client=7.1.0=pyhd3eb1b0_0
jupyter_core=4.8.1=py36h06a4308_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kiwisolver=1.3.1=py36h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libidn2=2.3.2=h7f8727e_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py36h06a4308_0
markupsafe=2.0.1=pypi_0
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
multidict=5.1.0=py36h27cfd23_2
nbclient=0.5.9=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=hbbd107a_1
networkx=2.5.1=pypi_0
ninja=1.10.2=h5e70eb0_2
notebook=6.4.6=pypi_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=py36_0
opencv-python=4.5.4.60=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.1.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=pypi_0
pip=21.2.2=py36h06a4308_0
prometheus-client=0.12.0=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.17.2=py36h295c915_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyarrow=6.0.1=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.6.2=pypi_0
pygments=2.10.0=pyhd3eb1b0_0
pyjwt=2.1.0=py36h06a4308_0
pympler=1.0.1=pypi_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
pyrsistent=0.18.0=pypi_0
pysocks=1.7.1=py36h06a4308_0
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.7.1=py3.6_cuda10.2.89_cudnn7.6.5_0
pytz=2021.3=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.2.1=py36h295c915_1
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
regex=2021.11.10=pypi_0
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
sacremoses=0.0.47=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.24.2=py36ha9443f7_0
scikit-video=1.1.11=pyh24bf2e0_0
scipy=1.5.2=py36h0b6359f_0
semver=2.13.0=pypi_0
send2trash=1.8.0=pypi_0
setuptools=58.0.4=py36h06a4308_0
sip=4.19.8=py36hf484d3e_0
six=1.16.0=pyhd3eb1b0_0
smmap=5.0.0=pypi_0
soupsieve=2.3.1=pypi_0
sqlite=3.37.0=hc218d9a_0
streamlit=1.9.2=pypi_0
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.0=py36hca6d32c_0
tensorboard-plugin-wit=1.6.0=py_0
terminado=0.12.1=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2020.9.3=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.10.3=pypi_0
toml=0.10.2=pypi_0
toolz=0.11.2=pypi_0
torch-fidelity=0.3.0=pypi_0
torchvision=0.8.2=py36_cu102
tornado=6.1=py36h27cfd23_0
tqdm=4.62.3=pyhd3eb1b0_1
traitlets=4.3.3=py36h06a4308_0
transformers=4.15.0=pypi_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021.5=pypi_0
tzlocal=4.1=pypi_0
urllib3=1.26.7=pypi_0
validators=0.18.2=pypi_0
watchdog=2.1.6=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=pypi_0
wikiextractor=3.0.6=pypi_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
yarl=1.6.3=py36h27cfd23_0
zeromq=4.3.4=h2531618_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
